% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonFitP1.R, R/PersonFitProject1.R
\name{Personfit}
\alias{Personfit}
\title{Item Response Theory Study}
\usage{
Personfit(x, complete = FALSE, ...)

Personfit(x, complete = FALSE, ...)
}
\arguments{
\item{x}{matrix or data.frame of scored item responses, one row per person,
one column per item.}

\item{complete}{logical with default \code{FALSE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{...}{further arguments passed to or from other functions, including
\code{\link{glmer}}}

\item{ip}{vector or matrix of item parameters, with a in the first column, b
in the second, and c in the third.}

\item{theta}{vector of thetas over which to estimate the corresponding function.}

\item{x}{matrix or data.frame of scored item responses, one row per person,
one column per item.}

\item{complete}{logical with default \code{FALSE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{ip}{vector or matrix of item parameters, with a in the first column, b
in the second, and c in the third.}

\item{theta}{vector of thetas over which to estimate the corresponding function.}

\item{...}{further arguments passed to or from other functions, including
\code{\link{glmer}}}
}
\description{
Functions for estimating person fit.
Currently, only the 1 parameter Rasch model is supported.

Functions for estimating person fit.
Currently, only the 1 parameter Rasch model is supported.
}

